+++ internal/utils/http.go
@@ -14,23 +14,6 @@ func CheckRespStatus(resp *resty.Response) error {
 	return nil
 }
 
-func ParseCookie(cookie string) (map[string]string, error) {
-	results := make(map[string]string)
-	splited := strings.Split(cookie, ";")
-	for _, item := range splited {
-		if len(item) == 0 {
-			continue
-		}
-		item = strings.TrimSpace(item)
-		kv := strings.SplitN(item, "=", 2)
-		if len(kv) != 2 {
-			return nil, fmt.Errorf("len(kv) should be 2 but '%v'", item)
-		}
-		results[kv[0]] = kv[1]
-	}
-	return results, nil
-}
-
 type HttpStatusError struct {
 	Code int
 	Msg  string
@@ -46,4 +29,11 @@ func IsStatusCode(err error, code int) bool {
 		return false
 	}
 	return e.Code == code
+}
+
+func StripAvatarSuffix(url string) string {
+	url = strings.Replace(url, "_normal", "", 1)
+	url = strings.Replace(url, "_bigger", "", 1)
+	url = strings.Replace(url, "_mini", "", 1)
+	return url
 }